V2 initial work

+++ src/main/java/dev/stashy/soundcategories/CategoryLoader.java
@@ -5,15 +5,13 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-public interface CategoryLoader
-{
+public interface CategoryLoader {
     /**
      * Registers a SoundCategory and injects its reference to the field that has this annotation.
      */
     @Retention(RetentionPolicy.RUNTIME)
     @Target(ElementType.FIELD)
-    @interface Register
-    {
+    @interface Register {
         /**
          * The ID of the sound category - if omitted, will be automatically set from the field name.
          */
@@ -29,6 +27,6 @@ public interface CategoryLoader
         /**
          * Allows changing the default volume level if one has not yet been set.
          */
-        float defaultLevel() default 1.0f;
+        double defaultLevel() default 1.0;
     }
 }
